requirements.txt
python-dotenv==1.0.0
pymongo==4.3.3
streamlit==1.23.1
pandas==1.2.4
pandas-flavor==0.2.0
pandasql==0.7.3
plotly==5.14.1
Pillow==8.2.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
numpy==1.23.5
sklearn==0.0
pickleshare==0.7.5
joblib==1.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2